Separate storage models from the storage business logic
