+++ user_management_service/app/storage/UserRepository.scala
@@ -1,6 +1,7 @@
 package storage
 
 import exceptions.StorageException
+import storage.model.UserResource
 
 import scala.concurrent.Future
